# docs/swagger.yaml
basePath: /api/v1
definitions:
  handler.CarbonResponse:
    properties:
      by_provider:
        items:
          $ref: '#/definitions/handler.ProviderCarbon'
        type: array
      by_region:
        items:
          $ref: '#/definitions/handler.RegionCarbon'
        type: array
    type: object
  handler.CleanupPreviewDTO:
    properties:
      action:
        example: delete
        type: string
      count:
        example: 5
        type: integer
      estimated_carbon_savings:
        example: 35.5
        type: number
      estimated_monthly_savings:
        example: 250
        type: number
      resources:
        items:
          $ref: '#/definitions/handler.ResourceDTO'
        type: array
    type: object
...
  handler.ErrorResponse:
    properties:
      error:
        example: invalid request
        type: string
    type: object
  handler.ExecuteCleanupRequest:
    properties:
      action:
        enum:
        - delete
        - stop
        - tag
        - notify
        example: delete
        type: string
      dry_run:
        example: false
        type: boolean
      organization_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      resource_ids:
        example:
        - 550e8400-e29b-41d4-a716-446655440001
        - 550e8400-e29b-41d4-a716-446655440002
        items:
          type: string
        minItems: 1
        type: array
    required:
    - action
    - organization_id
    - resource_ids
    type: object
  handler.ExecuteCleanupResponse:
    properties:
      dry_run:
        example: false
        type: boolean
      message:
        example: cleanup task queued
        type: string
      task_id:
        example: task_12345
        type: string
    type: object
...
  handler.ProviderCarbon:
    properties:
      carbon_kg:
        example: 450.25
        type: number
      potential_savings_kg:
        example: 75.5
        type: number
      provider:
        example: aws
        type: string
    type: object
  handler.ProviderSavings:
    properties:
      monthly_cost:
        example: 1500
        type: number
      potential_savings:
        example: 250
        type: number
      provider:
        example: aws
        type: string
      unused_count:
        example: 25
        type: integer
    type: object
...
  handler.RegionCarbon:
    properties:
      carbon_kg:
        example: 250
        type: number
      region:
        example: us-east-1
        type: string
    type: object
  handler.ResourceDTO:
    properties:
      carbon_footprint_kg:
        example: 12.5
        type: number
      created_at:
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      last_seen_at:
        type: string
      monthly_cost:
        example: 45.5
        type: number
      name:
        example: my-instance
        type: string
      organization_id:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
      provider:
        enum:
        - aws
        - azure
        - gcp
        example: aws
        type: string
      region:
        example: us-east-1
        type: string
      resource_id:
        example: i-1234567890abcdef0
        type: string
      status:
        enum:
        - active
        - unused
        - deleted
        - excluded
        example: unused
        type: string
      tags:
        additionalProperties:
          type: string
        type: object
      type:
        example: ec2_instance
        type: string
      updated_at:
        type: string
    type: object
  handler.SavingsResponse:
    properties:
      by_provider:
        items:
          $ref: '#/definitions/handler.ProviderSavings'
        type: array
      by_resource_type:
        items:
          $ref: '#/definitions/handler.TypeSavings'
        type: array
    type: object
...
  handler.SummaryStats:
    properties:
      potential_carbon_savings_kg:
        example: 180.25
        type: number
      potential_monthly_savings:
        example: 2500
        type: number
      total_carbon_kg:
        example: 1200.5
        type: number
      total_monthly_cost:
        example: 15000
        type: number
      total_resources:
        example: 500
        type: integer
      unused_resources:
        example: 75
        type: integer
    type: object
  handler.TypeSavings:
    properties:
      monthly_cost:
        example: 800
        type: number
      resource_type:
        example: ec2_instance
        type: string
      unused_count:
        example: 10
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@cloudsweep.io
    name: CloudSweep Support
    url: https://github.com/ermt0504/cloud-sweep
  description: API FinOps/GreenOps pour automatiser le nettoyage des ressources cloud
    inutilisees
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: CloudSweep API
  version: "1.0"
paths:
  /cleanup:
    post:
      consumes:
      - application/json
      description: Queue a cleanup operation for specified resources
      parameters:
      - description: Cleanup request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.ExecuteCleanupRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/handler.ExecuteCleanupResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Execute cleanup
      tags:
      - Cleanup
  /cleanup/preview:
    post:
      consumes:
      - application/json
      description: Preview what resources would be affected by a cleanup operation
      parameters:
      - description: Cleanup preview request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.ExecuteCleanupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.CleanupPreviewDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Preview cleanup
      tags:
      - Cleanup
  /dashboard/carbon:
    get:
      consumes:
      - application/json
      description: Get carbon footprint breakdown by provider and region
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.CarbonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Carbon footprint breakdown
      tags:
      - Dashboard
  /dashboard/savings:
    get:
      consumes:
      - application/json
      description: Get potential savings breakdown by provider and resource type
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.SavingsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Savings breakdown
      tags:
      - Dashboard
  /dashboard/summary:
    get:
      consumes:
      - application/json
      description: Get dashboard summary statistics including total resources, unused
        resources, costs and carbon footprint
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              $ref: '#/definitions/handler.SummaryStats'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Dashboard summary
      tags:
      - Dashboard
...
securityDefinitions:
  BearerAuth:
    description: Bearer token authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"